Organizing the struct of the projet
