--- GetOuthere/GetOuthere/Views/Challenges.swift
@@ -1,7 +1,0 @@
-import SwiftUI
-
-struct Challenges:View {
-    var body: some View {
-        /*@START_MENU_TOKEN@*//*@PLACEHOLDER=Hello, world!@*/Text("Hello, world!")/*@END_MENU_TOKEN@*/
-    }
-}
